requirements.txt
fastapi
uvicorn
langgraph
langchain-core
langchain-openai
langchain-aws
pandas
python-dotenv
pydantic
